Revert to absolute imports for generated Python2  code.
Maxime's opengeode and GUI code depend on the old way
of code generation. There appears to be no way to generate
python2 code that does relative imports, and doesn't
"infect" code that uses it to MIGRATE to relative imports.

Until we find the resources to move opengeode and the GUIs
from Python2 to Python2/relativeImports/andorPython3,
all the changes done to address relative imports have
to be reverted.

I hope I am wrong in this and there's some magic way to
mix and match relative-non-relative imports. But I don't
believe so - new code can of course be created to use
relative imports from the start, but there's 10 years
legacy code (1000s of lines) that use what DMT generates.

For people (like Konrad) that need Python2 relative imports
and/or write in Python3, I created a branch that has the
appropriate code:

    generatedCodeWithRelativeImportsForPy2andPy3

Sigh.