Augmenting the Simulink mapper for CoRA/Bambu needs
This commit acts as an example, demonstrating how the TASTE
abstract syntax trees can be used to emit the necessary
Bambu "bridge". More specifically, it demonstrates generating
the "prototype" part of the Bambu bridge function - and the
body of the function can be made in a very similar manner
(with another recursive function computing the necessary
lines).