Rename the ACN part in asn2aadlplus
in DataView.aadl, there was a part referencing all ACN files, however
the AADL identifier was based on the file name. In the case of several
ACN files with the same name but in different folders, we could have
duplicate identifiers in the AADL file.
This patch makes use of the module name instead of the filename for the
identifier. The module name must be unique.